src/requirements.txt
Django==1.11.5
django-crispy-forms==1.6.1
gunicorn==19.7.1
olefile==0.44
Pillow==4.1.1
psycopg2==2.7.1
pytz==2017.2

# Background Workers
redis==2.10.5
celery==3.1.23

# REST
djangorestframework==3.6.4
Markdown==2.6.8
djangorestframework-jwt==1.10
coreapi==2.3.0
Pygments==2.2.0
django-rest-swagger==2.0.5

# Stripe
stripe==1.65.0